shorcut to avoid unecessary work when located in a non-isolated environment
--HG--
branch : distribute
extra : rebase_source : 6b093a48e2647923cf36d1dd7ea74def0a961dc6

+++ tests/install_test.py
@@ -64,6 +64,7 @@ try:
     else:
         print '**** Test failed, please send me the output at tarek@ziade.org'
 finally:
-    os.remove(script_name)
+    if os.path.exists(script_name):
+        os.remove(script_name)
     os.remove('bootstrap.py')
